Refactor to implement GH ClientManager
Refactors the issue comment handler in order to implement a GH
ClientManager struct as a management wrapper for functions that interact
with GH APIs for no-phi-ai purposes.

Attempts to set a label on each issue based on the results of an AI
scan, as triggered by a given issue comment event.