Create a skeleton for running the app in cli mode
Initial setup for allowing the app to run in "cli" mode or "server"
mode, where "cli" mode is useful to allow for easier testing of new
code and features that do not fit neatly into the GitHub App flow.

Adds a new "pkg/client/no-git" package (aka "nogit") with a GitManager
struct and begins to add methods for working with (i.e. cloning) git
repositories.