Remove obsolete pkg/scanner
Adds the code required to finally support the removal of the now
obsolete "scanner" (v1) package. Adds translation from rrr.Request
to az.Document and from az.DocumentResponse to rrr.Response, along
with conversions for all other data structures.

Work-in-progress as the v2 Scanner does not currently guarantee that
it will wait for any responses, which becomes a bigger problem when
responses are not returned immediately, as with dry-run detector.